refactor: 로그인  UI 코드 정리
