+++ scripts/reset-project.js
