+++ src/features/signup/index.tsx
@@ -1,0 +1,8 @@
+import { Logo, LoginForm } from "./ui";
+
+const Signup = {
+  Logo, 
+  LoginForm,
+};
+
+export default Signup;
