+++ src/features/signup/ui/index.tsx
@@ -1,0 +1,2 @@
+export { default as Logo } from './logo';
+export { default as LoginForm } from './login-form';
